% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ebicat37}
\alias{ebicat37}
\title{EBI GWAS catalog (GRCh37)}
\format{An object of class \code{gwaswloc} of length 22688.}
\usage{
ebicat37
}
\description{
EBI GWAS catalog (GRCh37) from package `gwascat`
}
\keyword{datasets}
