% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_list}
\alias{pop_list}
\title{Population list}
\format{An object of class \code{list} of length 32.}
\usage{
pop_list
}
\description{
List to convert 1KGP populations to phase 1 superpopulation for haploR
}
\keyword{datasets}
