% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risklocienrich.R
\name{risk_loci_enrichment_test}
\alias{risk_loci_enrichment_test}
\title{Risk Loci Enrichment Test}
\usage{
risk_loci_enrichment_test(traits, regions, population = "EUR",
  ld_threshold = 0.5, ntimes = 50)
}
\arguments{
\item{traits}{character, list of disease traits}

\item{regions}{GRanges, regions to test}

\item{population}{1KGP population}

\item{ld_threshold}{LD threshold for definition of locus boundaries}

\item{ntimes}{number of permutations}
}
\value{
\describe{
  \item{perm_test}{result of the permutation test}
  \item{loci}{GRanges object containing loci}
  \item{regions}{GRanges object containing regions}
}
}
\description{
test regions for enrichment with disease risk loci
}
