% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risklocienrich.R
\name{setdiff_granges}
\alias{setdiff_granges}
\title{Set Difference of GRanges}
\usage{
setdiff_granges(x, ignore_strand = FALSE)
}
\arguments{
\item{x}{list of GRanges}

\item{ignore_strand}{ignore strand if TRUE}
}
\value{
GRanges, the set difference
}
\description{
Set Difference of GRanges
}
