% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risklocienrich.R
\name{snps_per_locus}
\alias{snps_per_locus}
\title{SNPs per locus}
\usage{
snps_per_locus(trait, ld_threshold = 0.5, population = "EUR")
}
\arguments{
\item{trait}{character, trait to consider}

\item{ld_threshold}{numeric, LD threshold for inclusion in a locus}

\item{population}{1KGP population for LD reference}
}
\value{
list of data frames giving haploR output per locus
}
\description{
Collect SNPs at trait associated loci based on LD with lead SNP
}
