% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risklocienrich.R
\name{snps_to_boundaries}
\alias{snps_to_boundaries}
\title{SNPS to boundaries}
\usage{
snps_to_boundaries(snps, ld_threshold = 0.5, population = "EUR")
}
\arguments{
\item{snps}{vector of SNPs at a locus}

\item{population}{1KGP population for LD reference}
}
\value{
\describe{
  \item{chr}{chromosome of the locus}
  \item{start}{start position of the locus}
  \item{end}{end position of the locus}
}
}
\description{
conver a vector of SNPS to boundaries
}
