% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valid_chromosomes}
\alias{valid_chromosomes}
\title{Valid chromosome strings}
\format{A character vector with 44 elements}
\usage{
valid_chromosomes
}
\description{
Vector of valid chromosome strings
}
\keyword{datasets}
